* patch-1:
  Fix number parsing of @Scheduled attributes

+++ spring-context/src/main/java/org/springframework/scheduling/annotation/ScheduledAnnotationBeanPostProcessor.java
@@ -75,6 +75,7 @@ import org.springframework.util.StringValueResolver;
  * @author Mark Fisher
  * @author Juergen Hoeller
  * @author Chris Beams
+ * @author Elizabeth Chatman
  * @since 3.0
  * @see Scheduled
  * @see EnableScheduling
@@ -295,7 +296,7 @@ public class ScheduledAnnotationBeanPostProcessor implements BeanPostProcessor, 
 					initialDelayString = this.embeddedValueResolver.resolveStringValue(initialDelayString);
 				}
 				try {
-					initialDelay = Integer.parseInt(initialDelayString);
+					initialDelay = Long.parseLong(initialDelayString);
 				}
 				catch (NumberFormatException ex) {
 					throw new IllegalArgumentException(
@@ -343,7 +344,7 @@ public class ScheduledAnnotationBeanPostProcessor implements BeanPostProcessor, 
 					fixedDelayString = this.embeddedValueResolver.resolveStringValue(fixedDelayString);
 				}
 				try {
-					fixedDelay = Integer.parseInt(fixedDelayString);
+					fixedDelay = Long.parseLong(fixedDelayString);
 				}
 				catch (NumberFormatException ex) {
 					throw new IllegalArgumentException(
@@ -367,7 +368,7 @@ public class ScheduledAnnotationBeanPostProcessor implements BeanPostProcessor, 
 					fixedRateString = this.embeddedValueResolver.resolveStringValue(fixedRateString);
 				}
 				try {
-					fixedRate = Integer.parseInt(fixedRateString);
+					fixedRate = Long.parseLong(fixedRateString);
 				}
 				catch (NumberFormatException ex) {
 					throw new IllegalArgumentException(
